refactor: reorganize admin chart routes
