+++ README.md
@@ -1,10 +1,10 @@
 # ACP+Charts now
-Current version: 0.0.53
+Current version: 0.0.54
 
 - Minimal React + Vite app with basic routing
 - Public pages: home and English release notes
 - Admin login at `/admin/login` using env credentials
-- Admin access to dashboard, analytics graphs, and UI demos of 30 forms and 30 hashtags
+- Admin access to charts dashboard, analytics graphs, and UI demos of 30 forms and 30 hashtags
 - Logout at `/admin/logout` with redirect for unauthenticated routes
 - Return to originally requested admin page after login
 - Public pages use a collapsible sidebar with icon tooltips and home link
@@ -13,10 +13,9 @@ Current version: 0.0.53
 - Admin pages display "Subpages:" with a full URL tree when subpages exist
 - Navigation sidebars derive from the same tree; admin lists all URLs flat (no nested lists), public lists non-admin URLs
 - User and admin sidebars highlight the active link
-- Admin charts with 20 Recharts examples for users data
-- Admin charts with 21 Chart.js examples for users data
-- Admin dashboard with mini charts for growth, engagement, reliability, and revenue
-- Detailed analytics pages at `/admin/graph/*` for growth, engagement, reliability, and revenue
+- Charts dashboard at `/admin/charts` with mini charts for growth, engagement, reliability, and revenue
+- Detailed analytics pages at `/admin/charts/*` for growth, engagement, reliability, and revenue
+- Chart.js users charts at `/admin/ui/charts`
 
 # ACP+Charts сomming soon
 ACP+Charts will grow into an admin dashboard that visualizes application metrics with interactive charts. Administrators will be able to monitor key indicators, manage data, and explore analytics through a responsive web interface built with React and Vite. The repository currently includes placeholder pages while chart components are under active development.
@@ -44,8 +43,8 @@ _Only this section of the readme can be maintained using Russian language_
   - [x] 3.3 Исправить невидимые графики на /admin/charts/users/recharts.
   - [x] 3.4 Пронумеровать варианты графиков в заголовках на /admin/charts/users/recharts.
   - [x] 3.5 Создать страницу /admin/charts/users/explain и описать ключи из users.json.
-  - [x] 3.6 Добавить страницу /admin/dashboard с мини-графиками.
-  - [x] 3.7 Реализовать страницы /admin/graph/growth, /admin/graph/engagement, /admin/graph/reliability, /admin/graph/revenue с реальными графиками и подписями.
+  - [x] 3.6 Добавить страницу /admin/charts с мини-графиками.
+  - [x] 3.7 Реализовать страницы /admin/charts/growth, /admin/charts/engagement, /admin/charts/reliability, /admin/charts/revenue с реальными графиками и подписями.
 
 6. Удобства
  - [x] 6.1 Изучить release-notes-howto.md. Создать файл release-notes.json и начать его вести. Указать в readme правила по ведению release-notes.json для каждого раза.
