+++ release-notes.json
@@ -1371,6 +1371,28 @@
           "scope": "analytics"
         }
       ]
+    },
+    {
+      "version": "0.0.54",
+      "date": "2025-08-29",
+      "time": "17:02:00",
+      "timezone": "Asia/Bishkek",
+      "changes": [
+        {
+          "description": "Reorganized admin chart routes and cleaned up sidebar",
+          "weight": 60,
+          "type": "refactor",
+          "scope": "admin-navigation"
+        }
+      ],
+      "changes-ru": [
+        {
+          "description": "Реорганизованы маршруты графиков админа и очищено меню",
+          "weight": 60,
+          "type": "refactor",
+          "scope": "admin-navigation"
+        }
+      ]
     }
   ],
   "releases": [
@@ -2633,6 +2655,28 @@
           "weight": 40,
           "type": "fix",
           "scope": "analytics"
+        }
+      ]
+    },
+    {
+      "version": "0.0.54",
+      "date": "2025-08-29",
+      "time": "17:02:00",
+      "timezone": "Asia/Bishkek",
+      "changes": [
+        {
+          "description": "Reorganized admin chart routes and cleaned up sidebar",
+          "weight": 60,
+          "type": "refactor",
+          "scope": "admin-navigation"
+        }
+      ],
+      "changes-ru": [
+        {
+          "description": "Реорганизованы маршруты графиков админа и очищено меню",
+          "weight": 60,
+          "type": "refactor",
+          "scope": "admin-navigation"
         }
       ]
     }
