--- src/admin/pages/adminGraphPage.jsx
@@ -1,17 +1,0 @@
-import { useEffect } from 'react'
-import { Outlet, useOutlet } from 'react-router-dom'
-import AuthMessage from '../app/authMessage.jsx'
-
-export default function AdminGraphPage() {
-  const title = 'Admin Graph Page'
-  const fullTitle = `${title} | Admin Control Panel | ACPC`
-  const outlet = useOutlet()
-  useEffect(() => { document.title = fullTitle }, [fullTitle])
-  if (outlet) return outlet
-  return (
-    <div>
-      <h1>{fullTitle}</h1>
-      <AuthMessage />
-    </div>
-  )
-}
